Permitir listar as finanças por mês

+++ FinancesApp.API/Models/Finance.cs
@@ -7,7 +7,7 @@ public class Finance
 {
     [Key]
     public int Id { get; private set; }
-    public int Value { get; private set; }
+    public decimal Value { get; private set; }
     public string Name { get; private set; }
     public string Description { get; private set; }
     public TransactionType TransactionType { get; private set; }
